@@ -47,9 +47,10 @@
   "devDependencies": {
     "@types/node": "^14.0.10",
     "@types/webpack": "^4.41.12",
-    "vue": "^2.6.8",
-    "vue-loader": "^15.7.0",
-    "vue-template-compiler": "^2.6.8"
+    "@vue/compiler-sfc": "^3.0.0-beta.14",
+    "vue": "^3.0.0-beta.14",
+    "vue-cli-plugin-vue-next": "^0.1.3",
+    "vue-loader": "^16.0.0-beta.3"
   },
   "peerDependencies": {
     "@babel/core": "*",
@@ -58,7 +59,7 @@
     "react": "*",
     "react-dom": "*",
     "ts-loader": "^6.2.2",
-    "vue": "^2.6.8",
+    "vue": "^3.0.0-beta.14",
     "vue-loader": "^15.7.0",
     "vue-template-compiler": "^2.6.8"
   },
