@@ -1,58 +1,79 @@
-import dedent from 'ts-dedent';
-import Vue from 'vue';
-import { RenderContext } from './types';
+// @ts-nocheck
+import dedent from 'ts-dedent'
+import { createApp, h, markRaw,ref } from 'vue'
+import { RenderContext } from './types'
 
-export const COMPONENT = 'STORYBOOK_COMPONENT';
-export const VALUES = 'STORYBOOK_VALUES';
+export const COMPONENT = 'STORYBOOK_COMPONENT'
+export const VALUES = 'STORYBOOK_VALUES'
 
-const root = new Vue({
-  data() {
+let mounted = false
+const activeComponent = ref(null)
+
+const root = createApp({
+  setup() {
     return {
-      [COMPONENT]: undefined,
-      [VALUES]: {},
-    };
+      activeComponent: ref(null)
+    }
   },
-  render(h) {
-    const children = this[COMPONENT] ? [h(this[COMPONENT])] : undefined;
-    return h('div', { attrs: { id: 'root' } }, children);
-  },
-});
+  render() {
 
-export default function render({
-  storyFn,
-  kind,
-  name,
-  showMain,
-  showError,
-  showException,
-  forceRender,
-}: RenderContext) {
-  Vue.config.errorHandler = showException;
+    const children = activeComponent.value ? [ h(activeComponent.value) ] : undefined
 
-  const element = storyFn();
+    return h('div', { attrs: { id: 'root' } }, children)
+
+  }
+})
+
+// const root = new Vue({
+//   data() {
+//     return {
+//       [ COMPONENT ]: undefined,
+//       [ VALUES ]: {}
+//     }
+//   },
+//   render(h) {
+//     const children = this[ COMPONENT ] ? [ h(this[ COMPONENT ]) ] : undefined
+//     return h('div', { attrs: { id: 'root' } }, children)
+//   }
+// })
+
+export default function render({ storyFn, kind, name, showMain, showError, showException, forceRender }: RenderContext) {
+
+  root.config.errorHandler = showException
+
+  const element = storyFn()
 
   if (!element) {
     showError({
-      title: `Expecting a Vue component from the story: "${name}" of "${kind}".`,
+      title: `Expecting a Vue component from the story: "${ name }" of "${ kind }".`,
       description: dedent`
         Did you forget to return the Vue component from the story?
         Use "() => ({ template: '<my-comp></my-comp>' })" or "() => ({ components: MyComp, template: '<my-comp></my-comp>' })" when defining the story.
-      `,
-    });
-    return;
+      `
+    })
+    return
   }
 
-  showMain();
+  showMain()
 
   // at component creation || refresh by HMR
-  if (!root[COMPONENT] || !forceRender) {
-    root[COMPONENT] = element;
+  if (!activeComponent.value || !forceRender) {
+
+    activeComponent.value = element
+
   }
 
-  // @ts-ignore https://github.com/storybookjs/storybook/pull/7578#discussion_r307986139
-  root[VALUES] = element.options[VALUES];
+  //
+  // // @ts-ignore https://github.com/storybookjs/storybook/pull/7578#discussion_r307986139
+  // root[ VALUES ] = element.options[ VALUES ]
 
-  if (!root.$el) {
-    root.$mount('#root');
+  // if (!root.$el) {
+  //   root.$mount('#root')
+  // }
+
+  if (!mounted) {
+    root.mount('#root')
+    mounted = true
   }
+
 }
