@@ -1,21 +1,15 @@
+// @ts-nocheck
 /* eslint-disable prefer-destructuring */
-import Vue, { VueConstructor, ComponentOptions } from 'vue';
-import { start } from '@storybook/core/client';
-import {
-  ClientStoryApi,
-  StoryFn,
-  DecoratorFunction,
-  StoryContext,
-  Loadable,
-} from '@storybook/addons';
+import { ClientStoryApi, DecoratorFunction, Loadable, StoryContext, StoryFn } from '@storybook/addons'
+import { start } from '@storybook/core/client'
+import { ComponentOptions, VueConstructor } from 'vue'
 
-import './globals';
-import { IStorybookSection, StoryFnVueReturnType } from './types';
+import './globals'
 
-import render, { VALUES } from './render';
-import { extractProps } from './util';
+import render from './render'
+import { IStorybookSection, StoryFnVueReturnType } from './types'
 
-export const WRAPS = 'STORYBOOK_WRAPS';
+export const WRAPS = 'STORYBOOK_WRAPS'
 
 function prepare(
   rawStory: StoryFnVueReturnType,
@@ -35,32 +29,33 @@ function prepare(
   // eslint-disable-next-line no-underscore-dangle
   if (!story._isVue) {
     if (innerStory) {
-      story.components = { ...(story.components || {}), story: innerStory };
+      story.components = { ...(story.components || {}), story: innerStory }
     }
-    story = Vue.extend(story);
     // @ts-ignore // https://github.com/storybookjs/storybook/pull/7578#discussion_r307984824
-  } else if (story.options[WRAPS]) {
-    return story as VueConstructor;
+  } else if (story.options[ WRAPS ]) {
+    return story as VueConstructor
   }
 
-  return Vue.extend({
-    // @ts-ignore // https://github.com/storybookjs/storybook/pull/7578#discussion_r307985279
-    [WRAPS]: story,
-    // @ts-ignore // https://github.com/storybookjs/storybook/pull/7578#discussion_r307984824
-    [VALUES]: { ...(innerStory ? innerStory.options[VALUES] : {}), ...extractProps(story) },
-    functional: true,
-    render(h, { data, parent, children }) {
-      return h(
-        story,
-        {
-          ...data,
-          // @ts-ignore // https://github.com/storybookjs/storybook/pull/7578#discussion_r307986196
-          props: { ...(data.props || {}), ...parent.$root[VALUES] },
-        },
-        children
-      );
-    },
-  });
+  return story
+
+  // return Vue.extend({
+  //   // @ts-ignore // https://github.com/storybookjs/storybook/pull/7578#discussion_r307985279
+  //   [ WRAPS ]: story,
+  //   // @ts-ignore // https://github.com/storybookjs/storybook/pull/7578#discussion_r307984824
+  //   [ VALUES ]: { ...(innerStory ? innerStory.options[ VALUES ] : {}), ...extractProps(story) },
+  //   functional: true,
+  //   render(h, { data, parent, children }) {
+  //     return h(
+  //       story,
+  //       {
+  //         ...data,
+  //         // @ts-ignore // https://github.com/storybookjs/storybook/pull/7578#discussion_r307986196
+  //         props: { ...(data.props || {}), ...parent.$root[ VALUES ] }
+  //       },
+  //       children
+  //     )
+  //   }
+  // })
 }
 
 const defaultContext: StoryContext = {
